+++ order-inventory/src/main/java/com/spring/order_inventory/repository/InventoryRepository.java
@@ -3,5 +3,5 @@ package com.spring.order_inventory.repository;
 import com.spring.order_inventory.entity.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<Inventory,Long> {
+public interface InventoryRepository extends JpaRepository<Inventory,Integer> {
 }
